backend/requirements.txt
Flask==1.1.2
paramiko==2.7.2
PyYAML==5.3.1
Pebble==4.6.0
